Allows a react environment inside a browser
When you run a RageMP react project in a browser, `mp` will be undefined. These changes check whether or not `mp` exists before use without having to change your React codebase.

This change is tailored for my project. I don't think the problem it solves is one that a lot of RageMP developers will encounter.